Refactor constants.ts into digitalTwinConfig/settingsUtil.ts and digitalTwinConfig/constants.ts
