+++ client/src/preview/route/digitaltwins/DigitalTwinsPreview.tsx
@@ -6,7 +6,7 @@ import Layout from 'page/Layout';
 import TabComponent from 'components/tab/TabComponent';
 import { TabData } from 'components/tab/subcomponents/TabRender';
 import AssetBoard from 'preview/components/asset/AssetBoard';
-import { defaultFiles } from 'model/backend/gitlab/constants';
+import { defaultFiles } from 'model/backend/gitlab/digitalTwinConfig/constants';
 import { addOrUpdateFile } from 'preview/store/file.slice';
 import tabs from './DigitalTwinTabDataPreview';
 import CreatePage from './create/CreatePage';
