+++ client/src/preview/util/digitalTwinUtils.ts
@@ -6,7 +6,8 @@ import {
   ProjectId,
 } from 'model/backend/gitlab/UtilityInterfaces';
 import { Asset } from 'preview/components/asset/Asset';
-import { AssetTypes, getDTDirectory } from 'model/backend/gitlab/constants';
+import { AssetTypes } from 'model/backend/gitlab/digitalTwinConfig/constants';
+import {  getDTDirectory } from 'model/backend/gitlab/digitalTwinConfig/settingsUtility';
 import GitlabInstance from 'model/backend/gitlab/instance';
 import { ExecutionStatus } from 'model/backend/gitlab/types/executionHistory';
 import DigitalTwin from './digitalTwin';
