+++ client/test/preview/unit/routes/digitaltwins/create/ConfirmDeleteDialog.test.tsx
@@ -6,7 +6,7 @@ import {
   removeAllCreationFiles,
   addOrUpdateFile,
 } from 'preview/store/file.slice';
-import { defaultFiles } from 'model/backend/gitlab/constants';
+import { defaultFiles } from 'model/backend/gitlab/digitalTwinConfig/constants';
 
 jest.mock('react-redux', () => ({
   ...jest.requireActual('react-redux'),
