+++ client/test/preview/unit/util/libraryAsset.test.ts
@@ -1,7 +1,8 @@
 import LibraryAsset, { getLibrarySubfolders } from 'preview/util/libraryAsset';
 import { BackendInterface } from 'model/backend/gitlab/UtilityInterfaces';
 import LibraryManager from 'preview/util/libraryManager';
-import { AssetTypes, getGroupName } from 'model/backend/gitlab/constants';
+import { AssetTypes } from 'model/backend/gitlab/digitalTwinConfig/constants';
+import {  getGroupName } from 'model/backend/gitlab/digitalTwinConfig/settingsUtility';
 
 jest.mock('preview/util/libraryManager');
